frontend/requirements.txt
# Frontend Dependencies

# Streamlit
streamlit==1.31.0

# Image processing
Pillow==10.2.0

# HTTP requests
requests==2.31.0
